Move restore audit code to different file
